move v1 cluster ssh client to test
